refactor: 책 상세 페이지 컴포넌트 훅으로 로직 분리
