+++ FE/src/components/Detail/LibrariesDisplay/LibrariesDisplay.tsx
@@ -2,13 +2,9 @@ import { EnvironmentOutlined, PhoneOutlined } from "@ant-design/icons";
 import { Heading } from "@components/Common";
 import KakaoMap from "@components/Common/KakaoMap/KakaoMap";
 import LoadingSpin from "@components/Common/Spin/Spin";
-import { message } from "antd";
-import { useEffect, useRef, useState } from "react";
 import styled, { css } from "styled-components";
 
-import { DEFAULT_INDEX } from "@/constants";
-import useCheckLoanStatus from "@/hooks/Queries/Detail/useCheckLoanStatus";
-import useLibraryOpenAPIQuery from "@/hooks/Queries/useLibraryOpenAPIQuery";
+import { useLibrariesDisplayLogic } from "@/hooks/BookDetail/useLibrariesDisplayLogic";
 import { LibrariesType } from "@/types/libraryType";
 interface LibrariesDisplayProps {
     isbn13: string;
@@ -21,51 +17,17 @@ const LibrariesDisplay = ({
     regionCode,
     subRegionCode,
 }: LibrariesDisplayProps) => {
-    const loanRequestRef = useRef<{ libCode: string; isRequest: boolean }>({
-        libCode: "",
-        isRequest: false,
-    });
-    const [selectedLibrary, setSelectedLibrary] =
-        useState<LibrariesType | null>(null);
-    const { data, isLoading } = useLibraryOpenAPIQuery(
-        "libSrchByBook",
-        [regionCode, subRegionCode, isbn13],
-        `&isbn=${isbn13}&region=${regionCode}&dtl_region=${subRegionCode}`
-    );
 
-    const { mutate } = useCheckLoanStatus(isbn13);
+    const {
+        loanRequestRef,
+        selectedLibrary,
+        setSelectedLibrary,
+        libraries,
+        isLoading,
+        handleLoanStatusClick,
+    } = useLibrariesDisplayLogic(isbn13, regionCode, subRegionCode);
 
-    useEffect(() => {
-        if (!isLoading && data.response.libs.length > 0) {
-            setSelectedLibrary(data.response.libs[DEFAULT_INDEX]);
-        }
-
-    }, [data, isLoading]);
-
-    
-    const handleLoanStatusClick = (libCode: string) => {
-        if (loanRequestRef.current.libCode)
-            return message.error("현재 대출 조회 중입니다.");
-        loanRequestRef.current.libCode = libCode;
-        loanRequestRef.current.isRequest = true;
-        
-        mutate(libCode, {
-            onSuccess: (data) => {
-                if (data.response.result.loanAvailable === "Y")
-                    return message.success("대출이 가능합니다!");
-                return message.warning("현재 대출이 불가능합니다..");
-            },
-            onSettled: () => {
-                loanRequestRef.current.libCode = "";
-                loanRequestRef.current.isRequest = false;
-            },
-        });
-    };
-
-    
-    if (isLoading) return <LoadingSpin/>;
-    
-    const libraries: LibrariesType[] = data.response.libs;
+    if (isLoading) return <LoadingSpin />;
     
     return (
         <>
@@ -77,7 +39,7 @@ const LibrariesDisplay = ({
                         <div>다른 지역을 선택해보세요!</div>
                     </NoLibrariesMessage>
                 ) : (
-                    libraries.map((item) => (
+                    libraries.map((item: LibrariesType) => (
                         <LibraryCard
                             key={item.lib.libCode}
                             $selected={
